+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 - Helm Release: fix username fetch option (https://github.com/pulumi/pulumi-kubernetes/pull/1824)
 
+- Fix CRD upgrades (https://github.com/pulumi/pulumi-kubernetes/pull/1819)
+
 ## 3.12.0 (December 7, 2021)
 
 - Add support for k8s v1.23.0. (https://github.com/pulumi/pulumi-kubernetes/pull/1681)